Move C headers
