+++ include/gauxc/c/functional.h
@@ -11,8 +11,8 @@
  */
 #pragma once
 
-#include <gauxc/types.h>
-#include <gauxc/status.h>
+#include <gauxc/c/types.h>
+#include <gauxc/c/status.h>
 
 #ifdef __cplusplus
 extern "C" {
