--- include/gauxc/gauxc_config.hpp
@@ -1,20 +1,0 @@
-/**
- * GauXC Copyright (c) 2020-2024, The Regents of the University of California,
- * through Lawrence Berkeley National Laboratory (subject to receipt of
- * any required approvals from the U.S. Dept. of Energy).
- *
- * (c) 2024-2025, Microsoft Corporation
- *
- * All rights reserved.
- *
- * See LICENSE.txt for details
- */
-#pragma once
-
-#include <gauxc/gauxc_config.h>
-
-#if defined(__CUDACC__) || defined(__HIPCC__)
-  #define HOST_DEVICE_ACCESSIBLE __host__ __device__
-#else
-  #define HOST_DEVICE_ACCESSIBLE
-#endif
