+++ include/gauxc/gauxc_config.hpp.in
@@ -1,0 +1,41 @@
+/**
+ * GauXC Copyright (c) 2020-2024, The Regents of the University of California,
+ * through Lawrence Berkeley National Laboratory (subject to receipt of
+ * any required approvals from the U.S. Dept. of Energy).
+ *
+ * (c) 2024-2025, Microsoft Corporation
+ *
+ * All rights reserved.
+ *
+ * See LICENSE.txt for details
+ */
+#pragma once
+
+#cmakedefine GAUXC_HAS_C
+#cmakedefine GAUXC_HAS_HOST
+#cmakedefine GAUXC_HAS_CUDA
+#cmakedefine GAUXC_HAS_HIP
+#cmakedefine GAUXC_HAS_MPI
+#cmakedefine GAUXC_HAS_MAGMA
+#cmakedefine GAUXC_HAS_NCCL
+#cmakedefine GAUXC_HAS_CUTLASS
+#cmakedefine GAUXC_HAS_GAU2GRID
+#cmakedefine GAUXC_HAS_HDF5
+#cmakedefine GAUXC_USE_FAST_RSQRT
+
+#ifdef GAUXC_HAS_HOST
+#cmakedefine GAUXC_CPU_XC_MAX_AM     @GAUXC_CPU_XC_MAX_AM@
+#cmakedefine GAUXC_CPU_SNLINK_MAX_AM @GAUXC_CPU_SNLINK_MAX_AM@
+#endif
+
+#cmakedefine GAUXC_HAS_DEVICE
+#ifdef GAUXC_HAS_DEVICE
+#cmakedefine GAUXC_GPU_XC_MAX_AM     @GAUXC_GPU_XC_MAX_AM@
+#cmakedefine GAUXC_GPU_SNLINK_MAX_AM @GAUXC_GPU_SNLINK_MAX_AM@
+#endif
+
+#if defined(__CUDACC__) || defined(__HIPCC__)
+  #define HOST_DEVICE_ACCESSIBLE __host__ __device__
+#else
+  #define HOST_DEVICE_ACCESSIBLE
+#endif
