+++ src/c-api/c_types.cxx
@@ -10,16 +10,16 @@
  * See LICENSE.txt for details
  */
 
-#include <gauxc/types.h>
-#include <gauxc/status.h>
-#include <gauxc/molecule.h>
-#include <gauxc/basisset.h>
-#include <gauxc/molgrid.h>
-#include <gauxc/runtime_environment.h>
-#include <gauxc/load_balancer.h>
-#include <gauxc/molecular_weights.h>
-#include <gauxc/functional.h>
-#include <gauxc/xc_integrator.h>
+#include <gauxc/c/types.h>
+#include <gauxc/c/status.h>
+#include <gauxc/c/molecule.h>
+#include <gauxc/c/basisset.h>
+#include <gauxc/c/molgrid.h>
+#include <gauxc/c/runtime_environment.h>
+#include <gauxc/c/load_balancer.h>
+#include <gauxc/c/molecular_weights.h>
+#include <gauxc/c/functional.h>
+#include <gauxc/c/xc_integrator.h>
 
 #include "c_status.hpp"
